refactor: rename `NpcsEventType` to `NpcEventType` and update references, restructure managers into `Managers` namespace
