+++ src/Extensions/NeoServer.Scripts.LuaJIT/Interfaces/INpcs.cs
@@ -4,7 +4,7 @@ namespace NeoServer.Scripts.LuaJIT.Interfaces;
 
 public interface INpcs
 {
-    void Add(string npcName, NpcsEventType eventType, LuaScriptInterface luaScriptInterface);
+    void Add(string npcName, NpcEventType eventType, LuaScriptInterface luaScriptInterface);
     bool LoadCallback(string npcName);
     NpcEvents GetEvents(string npcName);
     void Clear();
