+++ src/Extensions/NeoServer.Scripts.LuaJIT/IoC/Modules/LuaJITInjection.cs
@@ -2,6 +2,7 @@
 using NeoServer.Scripts.LuaJIT.Functions;
 using NeoServer.Scripts.LuaJIT.Functions.Interfaces;
 using NeoServer.Scripts.LuaJIT.Interfaces;
+using NeoServer.Scripts.LuaJIT.Managers;
 using NeoServer.Scripts.LuaJIT.Services;
 
 namespace NeoServer.Scripts.LuaJIT.IoC.Modules;
