+++ src/Extensions/NeoServer.Scripts.LuaJIT/LuaStartup.cs
@@ -1,6 +1,7 @@
 ﻿using LuaNET;
 using NeoServer.Scripts.LuaJIT.Functions.Interfaces;
 using NeoServer.Scripts.LuaJIT.Interfaces;
+using NeoServer.Scripts.LuaJIT.Managers;
 using NeoServer.Server.Configurations;
 using Serilog;
