+++ src/Extensions/NeoServer.Scripts.LuaJIT/Scripts.cs
@@ -3,6 +3,7 @@ using NeoServer.Domain.Spells;
 using NeoServer.Scripts.LuaJIT.Enums;
 using NeoServer.Scripts.LuaJIT.Enums.Config;
 using NeoServer.Scripts.LuaJIT.Interfaces;
+using NeoServer.Scripts.LuaJIT.Managers;
 using Serilog;
 
 namespace NeoServer.Scripts.LuaJIT;
