+++ src/Standalone/IoC/Modules/LuaInjection.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using NeoServer.Scripts.LuaJIT;
+using NeoServer.Scripts.LuaJIT.Managers;
 using NeoServer.Scripts.LuaJIT.ScriptServices;
 using NeoServer.Server.Common.Contracts.Scripts;
 using NeoServer.Server.Common.Contracts.Scripts.Services;
