+++ tests/NeoServer.Domain.Tests/Server/ScriptManagerTestBuilder.cs
@@ -2,6 +2,7 @@
 using NeoServer.Domain.Common.Contracts.Items;
 using NeoServer.Scripts.LuaJIT;
 using NeoServer.Scripts.LuaJIT.Interfaces;
+using NeoServer.Scripts.LuaJIT.Managers;
 using NeoServer.Server.Common.Contracts.Scripts;
 using NeoServer.Server.Common.Contracts.Scripts.Services;
 using NeoServer.Server.Configurations;
